requirements.txt
adbutils==0.8.1
airtest==1.1.3
apkutils2==1.0.0
certifi==2020.6.20
chardet==3.0.4
cigam==0.0.3
click==7.1.2
decorator==4.4.2
deprecation==2.1.0
facebook-wda==0.2.1
Flask==1.1.2
Flask-Cors==3.0.8
hrpc==1.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.19.0
opencv-contrib-python==3.4.2.17
packaging==20.4
Pillow==7.1.2
pocoui==1.0.79
py==1.9.0
pyelftools==0.26
pyparsing==2.4.7
python-xlib==0.27
pywinauto==0.6.3
requests==2.24.0
retry==0.9.2
six==1.15.0
tenacity==6.2.0
urllib3==1.25.9
websocket-client==0.48.0
Werkzeug==1.0.1
whichcraft==0.6.1
xmltodict==0.12.0
